+++ .env.sample
@@ -1,2 +1,2 @@
-PORT=8081
+PORT=
 CONNECTION_DB=
