--- index.js
@@ -1,2 +1,0 @@
-require('dotenv').config()
-require('./app/server')
